\name{dynSBevo Package}
\alias{dynSBevo Package}
\alias{dynSBevo}
\docType{package}
\title{Overview of the package dynSBevo}

\description{\packageDescription{dynSBevo}}

\details{
The main functions of the package are listed hereafter:
\itemize{
  \item \code{\link{dynSB}}: performs approximate maximum likelihood estimation of the parameters of the dynamic stochastic block model for longitudinal network data using VEM or EVEM algorithms. 
  \item \code{\link{dynSB_Sample}}: randomly sample a longitudinal network according to the dynamic stochastic block model.
}
}

\author{
Luca Brusa [aut, cre], Fulvia Pennoni [aut]

\emph{Maintainer}: Luca Brusa (luca.brusa@unimib.it)
}
